perf: prepare plugin package ownership once per phase
Keep capture and replacement reconciliation in separate synchronous resolver lifetimes. Preserve package identity, duplicate child ordering, conflict checks and independent result arrays while removing repeated preparation from internal callers.
