+++ src/plugins/installed-plugin-package-ownership.ts
@@ -1,4 +1,5 @@
 import path from "node:path";
+import type { Result } from "@openclaw/normalization-core/result";
 import { isPathInside } from "../infra/path-guards.js";
 import { resolveUserPath } from "../utils.js";
 import {
@@ -14,10 +15,10 @@ import { safeRealpathSync } from "./path-safety.js";
 function collectDuplicateInstallRecordOwners(
   index: InstalledPluginIndex,
   env: NodeJS.ProcessEnv,
+  realpathCache: Map<string, string>,
 ): Set<string> {
   const ownersByPath = new Map<string, string>();
   const duplicateOwners = new Set<string>();
-  const realpathCache = new Map<string, string>();
   for (const [installOwner, record] of Object.entries(index.installRecords)) {
     const rawPath = record.installPath?.trim() || record.sourcePath?.trim();
     if (!rawPath) {
@@ -50,13 +51,9 @@ export type InstalledPluginLifecycleOwnership =
       pluginIds: [];
     };
 
-type InstalledPluginPackageOwnershipResult =
-  | { ok: true; value: InstalledPluginPackageOwnership }
-  | { ok: false; error: string };
+type InstalledPluginPackageOwnershipResult = Result<InstalledPluginPackageOwnership, string>;
 
-type InstalledPluginLifecycleOwnershipResult =
-  | { ok: true; value: InstalledPluginLifecycleOwnership }
-  | { ok: false; error: string };
+type InstalledPluginLifecycleOwnershipResult = Result<InstalledPluginLifecycleOwnership, string>;
 
 function ownershipError(pluginId: string, detail: string): InstalledPluginPackageOwnershipResult {
   return {
@@ -67,107 +64,117 @@ function ownershipError(pluginId: string, detail: string): InstalledPluginPackag
   };
 }
 
-export function resolveInstalledPluginPackageOwnership(
+// Reuse only while the index and package paths are unchanged in one synchronous
+// phase. After yielding or replacing paths, create a fresh resolver.
+export function createInstalledPluginOwnershipResolver(
   index: InstalledPluginIndex,
-  pluginId: string,
   env: NodeJS.ProcessEnv = process.env,
-): InstalledPluginPackageOwnershipResult {
-  const target = index.plugins.find((entry) => entry.pluginId === pluginId);
-  if (target && isInstalledPluginIndexInstallOwnerAmbiguous(target)) {
-    return ownershipError(pluginId, "has ambiguous package ownership");
+) {
+  const targets = new Map<string, InstalledPluginIndex["plugins"][number]>();
+  const childrenByOwner = new Map<string, string[]>();
+  for (const entry of index.plugins) {
+    if (!targets.has(entry.pluginId)) {
+      targets.set(entry.pluginId, entry);
+    }
+    const installOwner = resolveInstalledPluginIndexInstallOwner(entry);
+    if (installOwner) {
+      const children = childrenByOwner.get(installOwner) ?? [];
+      children.push(entry.pluginId);
+      childrenByOwner.set(installOwner, children);
+    }
   }
-  const ownerFromTarget = target ? resolveInstalledPluginIndexInstallOwner(target) : undefined;
-  if (target && !ownerFromTarget) {
-    return ownershipError(pluginId, "has no authoritative package-owner metadata");
+  for (const children of childrenByOwner.values()) {
+    children.sort();
+  }
+  const realpathCache = new Map<string, string>();
+  let duplicateOwners: Set<string> | undefined;
+  const duplicates = () =>
+    (duplicateOwners ??= collectDuplicateInstallRecordOwners(index, env, realpathCache));
+  const unsafeOwners = new Map<string, boolean>();
+
+  function resolvePackage(pluginId: string): InstalledPluginPackageOwnershipResult {
+    const target = targets.get(pluginId);
+    if (target && isInstalledPluginIndexInstallOwnerAmbiguous(target)) {
+      return ownershipError(pluginId, "has ambiguous package ownership");
+    }
+    const ownerFromTarget = target ? resolveInstalledPluginIndexInstallOwner(target) : undefined;
+    if (target && !ownerFromTarget) {
+      return ownershipError(pluginId, "has no authoritative package-owner metadata");
+    }
+
+    const ownerFromRecord = Object.hasOwn(index.installRecords, pluginId) ? pluginId : undefined;
+    const installOwner = ownerFromTarget ?? ownerFromRecord;
+    if (!installOwner) {
+      return ownershipError(pluginId, "is not associated with a tracked package install");
+    }
+    if (ownerFromTarget && ownerFromRecord && ownerFromTarget !== ownerFromRecord) {
+      return ownershipError(pluginId, "matches conflicting package owners");
+    }
+    const installRecord = index.installRecords[installOwner];
+    if (!installRecord) {
+      return ownershipError(pluginId, `references missing package owner "${installOwner}"`);
+    }
+    if (duplicates().has(installOwner)) {
+      return ownershipError(pluginId, `shares package path ownership with "${installOwner}"`);
+    }
+
+    const [firstPluginId, ...remainingPluginIds] = childrenByOwner.get(installOwner) ?? [];
+    if (!firstPluginId) {
+      return ownershipError(
+        pluginId,
+        `package owner "${installOwner}" has no authoritative runtime child list`,
+      );
+    }
+    // Each result owns its child list; callers cannot mutate the prepared order.
+    const pluginIds: [string, ...string[]] = [firstPluginId, ...remainingPluginIds];
+    const hasUnsafePackageEntry =
+      unsafeOwners.get(installOwner) ??
+      index.plugins.some(
+        (entry) =>
+          installRecordPathMatchesPluginRoot(installRecord, entry.rootDir, env, realpathCache) &&
+          resolveInstalledPluginIndexInstallOwner(entry) !== installOwner,
+      );
+    unsafeOwners.set(installOwner, hasUnsafePackageEntry);
+    if (hasUnsafePackageEntry) {
+      return ownershipError(pluginId, `package owner "${installOwner}" has conflicting child rows`);
+    }
+    return {
+      ok: true,
+      value: { installOwner, installRecord, pluginIds },
+    };
   }
 
-  const ownerFromRecord = Object.hasOwn(index.installRecords, pluginId) ? pluginId : undefined;
-  const installOwner = ownerFromTarget ?? ownerFromRecord;
-  if (!installOwner) {
-    return ownershipError(pluginId, "is not associated with a tracked package install");
-  }
-  if (ownerFromTarget && ownerFromRecord && ownerFromTarget !== ownerFromRecord) {
-    return ownershipError(pluginId, "matches conflicting package owners");
-  }
-  const installRecord = index.installRecords[installOwner];
-  if (!installRecord) {
-    return ownershipError(pluginId, `references missing package owner "${installOwner}"`);
-  }
-  if (collectDuplicateInstallRecordOwners(index, env).has(installOwner)) {
-    return ownershipError(pluginId, `shares package path ownership with "${installOwner}"`);
-  }
-
-  const [firstPluginId, ...remainingPluginIds] = index.plugins
-    .filter(
+  function resolveLifecycle(pluginId: string): InstalledPluginLifecycleOwnershipResult {
+    const ownership = resolvePackage(pluginId);
+    if (ownership.ok) {
+      return { ok: true, value: { kind: "package", ...ownership.value } };
+    }
+    const installRecord = index.installRecords[pluginId];
+    if (
+      !Object.hasOwn(index.installRecords, pluginId) ||
+      !installRecord ||
+      duplicates().has(pluginId)
+    ) {
+      return ownership;
+    }
+    const hasConflictingEntry = index.plugins.some(
       (entry) =>
-        resolveInstalledPluginIndexInstallOwner(entry) === installOwner &&
-        !isInstalledPluginIndexInstallOwnerAmbiguous(entry),
-    )
-    .map((entry) => entry.pluginId)
-    .toSorted();
-  if (!firstPluginId) {
-    return ownershipError(
-      pluginId,
-      `package owner "${installOwner}" has no authoritative runtime child list`,
+        entry.pluginId === pluginId ||
+        installRecordPathMatchesPluginRoot(installRecord, entry.rootDir, env, realpathCache),
     );
+    if (hasConflictingEntry) {
+      return ownership;
+    }
+    // Cleanup and pre-update planning may act on an exact durable tombstone.
+    // Replacement reconciliation keeps using the strict package resolver.
+    return {
+      ok: true,
+      value: { kind: "orphan", installOwner: pluginId, installRecord, pluginIds: [] },
+    };
   }
-  const pluginIds: [string, ...string[]] = [firstPluginId, ...remainingPluginIds];
-  if (target && !pluginIds.includes(target.pluginId)) {
-    return ownershipError(pluginId, `does not belong to package owner "${installOwner}"`);
-  }
-
-  const realpathCache = new Map<string, string>();
-  const hasUnsafePackageEntry = index.plugins.some(
-    (entry) =>
-      installRecordPathMatchesPluginRoot(installRecord, entry.rootDir, env, realpathCache) &&
-      (isInstalledPluginIndexInstallOwnerAmbiguous(entry) ||
-        resolveInstalledPluginIndexInstallOwner(entry) !== installOwner),
-  );
-  if (hasUnsafePackageEntry) {
-    return ownershipError(pluginId, `package owner "${installOwner}" has conflicting child rows`);
-  }
-  return {
-    ok: true,
-    value: { installOwner, installRecord, pluginIds },
-  };
+  return { resolvePackage, resolveLifecycle };
 }
 
-export function resolveInstalledPluginLifecycleOwnership(
-  index: InstalledPluginIndex,
-  pluginId: string,
-  env: NodeJS.ProcessEnv = process.env,
-): InstalledPluginLifecycleOwnershipResult {
-  const ownership = resolveInstalledPluginPackageOwnership(index, pluginId, env);
-  if (ownership.ok) {
-    return { ok: true, value: { kind: "package", ...ownership.value } };
-  }
-  const installRecord = index.installRecords[pluginId];
-  if (
-    !Object.hasOwn(index.installRecords, pluginId) ||
-    !installRecord ||
-    collectDuplicateInstallRecordOwners(index, env).has(pluginId)
-  ) {
-    return ownership;
-  }
-  const realpathCache = new Map<string, string>();
-  const hasConflictingEntry = index.plugins.some(
-    (entry) =>
-      entry.pluginId === pluginId ||
-      installRecordPathMatchesPluginRoot(installRecord, entry.rootDir, env, realpathCache),
-  );
-  if (hasConflictingEntry) {
-    return ownership;
-  }
-  // Cleanup and pre-update planning may act on an exact durable tombstone.
-  // Replacement reconciliation keeps using the strict package resolver.
-  return {
-    ok: true,
-    value: { kind: "orphan", installOwner: pluginId, installRecord, pluginIds: [] },
-  };
-}
-
-// Share physical path facts only within one synchronous scan. A later lifecycle
-// check gets a fresh map so replaced package paths are observed again.
 function installRecordPathMatchesPluginRoot(
   record: InstalledPluginInstallRecordInfo,
   rootDir: string,
@@ -190,11 +197,11 @@ export function hasMissingInstalledPluginOwnerMetadata(
   index: InstalledPluginIndex,
   env: NodeJS.ProcessEnv = process.env,
 ): boolean {
-  if (collectDuplicateInstallRecordOwners(index, env).size > 0) {
+  const realpathCache = new Map<string, string>();
+  if (collectDuplicateInstallRecordOwners(index, env, realpathCache).size > 0) {
     return true;
   }
   const installRecords = Object.entries(index.installRecords);
-  const realpathCache = new Map<string, string>();
   // An orphaned owner record (for example, package code removed out of band) is
   // already closed by the lifecycle resolver. It must not make every unrelated
   // config read attempt an impossible registry migration with no discoverable rows.
