+++ src/plugins/manifest-registry-installed.test-helpers.ts
@@ -1,0 +1,32 @@
+import path from "node:path";
+import type { InstalledPluginIndex } from "./installed-plugin-index.js";
+
+export function createIndex(rootDir: string): InstalledPluginIndex {
+  return {
+    version: 1,
+    hostContractVersion: "2026.4.25",
+    compatRegistryVersion: "compat-v1",
+    migrationVersion: 1,
+    policyHash: "policy-v1",
+    generatedAtMs: 1777118400000,
+    installRecords: {},
+    plugins: [
+      {
+        pluginId: "installed",
+        manifestPath: path.join(rootDir, "openclaw.plugin.json"),
+        manifestHash: "manifest-hash",
+        source: path.join(rootDir, "index.ts"),
+        rootDir,
+        origin: "global",
+        enabled: true,
+        startup: {
+          sidecar: false,
+          memory: false,
+          agentHarnesses: [],
+        },
+        compat: [],
+      },
+    ],
+    diagnostics: [],
+  };
+}
